backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "portofino-portfolio-simulator"
version = "3.0"
description = "PortoFino Portfolio Simulator"
authors = [{ name = "52Sides", email = "everybody.knows.1152@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "yfinance>=0.2",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.0",
  "pydantic-settings",
  "sqlalchemy",
  "asyncpg",
  "psycopg2-binary",
  "redis",
  "celery",
  "pytest-celery",
  "kafka-python",
  "aiokafka",
  "python-dotenv",
  "httpx>=0.27",
  "loguru",
  "alembic",
  "tenacity",
  "pytest>=8.0",
  "pytest-cov",
  "pytest-asyncio>=0.23",
  "flake8",
  "black",
  "isort",
  "sqlalchemy[asyncio]>=2.0.25",
  "aiosqlite",
  "python-jose[cryptography]",
  "passlib",
  "pydantic[email]",
  "prometheus-fastapi-instrumentator>=6.0.0",
  "openpyxl",
  "argon2-cffi",
  "authlib>=1.2",
  "flower>=1.1.0",
  "python-multipart",
  "aiohttp",
  "itsdangerous>=2.1.2"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["api", "core", "db", "schemas", "services"]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "*/.venv/*"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = " --disable-warnings --cov=src --cov-branch --cov-report=term-missing"
testpaths = ["tests"]
markers = [
  "unit: isolated tests without external deps",
  "integration: tests using Postgres, Redis, Kafka, or API",
  "e2e: tests for pipelines"
]

